moving files and minor fix
